Refactoring / Renaming closes #63
