+++ apistar/commands/create_tables.py
@@ -5,7 +5,7 @@ def create_tables():
     """
     Create SQLAlchemy tables.
     """
-    from apistar.main import get_current_app
+    from apistar.cli import get_current_app
     from apistar.backends import SQLAlchemy
     app = get_current_app()
     db_backend = SQLAlchemy.build(settings=app.settings)
