+++ apistar/templating.py
@@ -4,7 +4,7 @@ from typing import List  # noqa
 import jinja2
 
 from apistar.exceptions import ConfigurationError
-from apistar.pipelines import ArgName
+from apistar.core import ArgName
 from apistar.settings import Settings
